--- UI/preview/pedalboard/ChordToMIDI.py
@@ -1,1 +1,0 @@
-def arms_to_MIDI(left_arm, right_arm, tempo):
